requirements.txt
absl-py==0.7.1
asn1crypto==0.24.0
astor==0.7.1
astroid==2.2.5
cachetools==3.1.0
certifi==2019.3.9
cffi==1.12.2
chardet==3.0.4
configparser==3.7.4
cryptography==2.6.1
defusedxml==0.6.0
Django==2.1.5
django-articles==2.4.1
django-ckeditor==5.6.1
django-classy-tags==0.8.0
django-cms==3.6.0
django-filer==1.4.4
django-formtools==2.1
django-js-asset==1.1.0
django-mptt==0.9.1
django-paypal==1.0.0
django-polymorphic==2.0.3
django-sekizai==0.10.0
django-taggit==1.1.0
django-treebeard==4.3
django-widget-tweaks==1.4.3
djangocms-admin-style==1.3.0
easy-thumbnails==2.5
gast==0.2.2
google-api-python-client==1.7.8
google-auth==1.6.3
google-auth-httplib2==0.0.3
googledrivedownloader==0.3
grpcio==1.20.0
h5py==2.9.0
html5lib==1.0.1
httplib2==0.12.1
idna==2.8
isort==4.3.16
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
lazy-object-proxy==1.3.1
Markdown==3.1
mccabe==0.6.1
mock==2.0.0
mysql-connector==2.1.6
mysql-connector-python==2.0.4
mysqlclient==1.4.1
numpy==1.16.3
oauthlib==3.0.1
pbr==5.1.3
pep8==1.7.1
Pillow==5.4.1
protobuf==3.7.1
pyasn1==0.4.5
pyasn1-modules==0.2.4
pycparser==2.19
PyJWT==1.7.1
pylint==2.3.1
pylint-django==2.0.9
pylint-plugin-utils==0.5
PyMySQL==0.9.3
python3-openid==3.1.0
pytz==2018.9
requests==2.21.0
requests-oauthlib==1.2.0
rsa==4.0
six==1.12.0
social-auth-core==3.1.0
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
typed-ast==1.3.1
Unidecode==1.0.23
uritemplate==3.0.0
urllib3==1.24.1
webencodings==0.5.1
Werkzeug==0.15.2
wrapt==1.11.1
